Align segments and batches with workflow endpoint ownership
